renamed package DukeObjects to list (see package naming standards)
