+++ src/main/java/command/AddExpense.java
@@ -1,8 +1,8 @@
 package command;
 
 
-import DukeObjects.Expense;
-import DukeObjects.ExpenseList;
+import list.Expense;
+import list.ExpenseList;
 import parser.CommandParams;
 import storage.Storage;
 import ui.Ui;
