+++ src/main/java/command/Command.java
@@ -1,6 +1,6 @@
 package command;
 
-import DukeObjects.ExpenseList;
+import list.ExpenseList;
 import parser.CommandParams;
 import storage.Storage;
 import ui.Ui;
