+++ src/main/java/command/ExitCommand.java
@@ -1,7 +1,6 @@
 package command;
 
-import DukeObjects.Expense;
-import DukeObjects.ExpenseList;
+import list.ExpenseList;
 import parser.CommandParams;
 import storage.Storage;
 import ui.Ui;
