+++ src/main/java/list/ExpenseList.java
@@ -1,7 +1,6 @@
-package DukeObjects;
+package list;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 public class ExpenseList{
